services/api_gateway/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic-settings
pyjwt
supabase
